add owner control

+++ app/Http/Livewire/ShowUsers.php
@@ -3,6 +3,8 @@
 namespace App\Http\Livewire;
 
 use App\Models\Office;
+use App\Models\User;
+use Illuminate\Support\Facades\Auth;
 use Livewire\Component;
 
 class ShowUsers extends Component
@@ -28,8 +30,35 @@ class ShowUsers extends Component
 
     public function deleteUser($user_id)
     {
+        if(!$this->current_office->isOwner(Auth::id())){
+            $this->notification()->error(
+                $title = 'User not deleted',
+                $description = 'You are not the owner of this office'
+            );
+            return; 
+        }
+        $user_to_delete = User::find($user_id);
+        if(!$user_to_delete){
+            $this->notification()->error(
+                $title = 'User not deleted',
+                $description = 'User do not exists'
+            );
+            return; 
+        }
+        if($this->current_office->isOwner($user_id)){
+            $this->notification()->error(
+                $title = 'Office not deleted',
+                $description = 'User is the owner of this office'
+            );
+            return; 
+        }
+        
         $this->current_office->users()->detach($user_id);
         $this->emit('refreshUser');
+        $this->notification()->success(
+            $title = 'User deleted successfully',
+            $description = 'You deleted the user successfully'
+        );
     }
 
     public function render()
